requirements.txt
scrapling
python-dotenv
